Reduce unnecessary dependencies
