+++ crates/holochain_sqlite/Cargo.toml
@@ -21,12 +21,14 @@ holo_hash = { path = "../holo_hash", version = "^0.3.0-beta-dev.6"}
 holochain_serialized_bytes = "=0.0.51"
 holochain_util = { version = "^0.2.0", path = "../holochain_util", features = ["backtrace"] }
 holochain_zome_types = { version = "^0.3.0-beta-dev.8", path = "../holochain_zome_types" }
-kitsune_p2p = { version = "^0.3.0-beta-dev.9", path = "../kitsune_p2p/kitsune_p2p" }
+kitsune_p2p_types = { version = "0.3.0-beta-dev.5", path = "../kitsune_p2p/types" }
+kitsune_p2p_dht_arc = { version = "0.3.0-beta-dev.3", path = "../kitsune_p2p/dht_arc" }
+kitsune_p2p_bin_data = { version = "0.3.0-beta-dev.4", path = "../kitsune_p2p/bin_data" }
+kitsune_p2p_dht = { version = "0.3.0-beta-dev.3", path = "../kitsune_p2p/dht" }
+kitsune_p2p_timestamp = { version = "0.2.0", path = "../kitsune_p2p/timestamp" }
 once_cell = "1.4.1"
-nanoid = "0.3.0"
 num_cpus = "1.13.0"
 parking_lot = "0.10"
-rand = "0.8.5"
 r2d2 = "0.8"
 r2d2_sqlite = { version = "0.1", package = "r2d2_sqlite_neonphog" }
 rmp-serde = "0.15"
@@ -57,6 +59,7 @@ rusqlite = { version = "0.29", features = [
 holochain_sqlite = { version = "0.3.0-beta-dev.8", path = ".", features = ["test_utils", "slow_tests"] }
 holochain_trace = { version = "^0.3.0-beta-dev.1", path = "../holochain_trace" }
 nanoid = "0.4.0"
+rand = "0.8.5"
 
 [build-dependencies]
 pretty_assertions = "0.7.2"
@@ -77,7 +80,9 @@ sqlite-encrypted = [
   "rusqlite/bundled-sqlcipher-vendored-openssl",
   "holo_hash/sqlite-encrypted",
   "holochain_zome_types/sqlite-encrypted",
-  "kitsune_p2p/sqlite-encrypted",
+  "kitsune_p2p_bin_data/sqlite-encrypted",
+  "kitsune_p2p_types/sqlite-encrypted",
+  "kitsune_p2p_dht_arc/sqlite-encrypted",
 ]
 
 # Compile SQLite from source rather than depending on a library
@@ -85,6 +90,8 @@ sqlite = [
   "rusqlite/bundled",
   "holo_hash/sqlite",
   "holochain_zome_types/sqlite",
-  "kitsune_p2p/sqlite",
+  "kitsune_p2p_bin_data/sqlite",
+  "kitsune_p2p_dht_arc/sqlite",
+  "kitsune_p2p_types/sqlite",
   "r2d2_sqlite/bundled",
 ]
