+++ crates/kitsune_p2p/kitsune_p2p/src/host_api.rs
@@ -15,7 +15,7 @@ use kitsune_p2p_types::{
     KOpData, KOpHash,
 };
 
-use crate::event::{GetAgentInfoSignedEvt, MetricRecord};
+use crate::event::GetAgentInfoSignedEvt;
 
 /// A boxed future result with dynamic error type
 pub type KitsuneHostResult<'a, T> =
@@ -130,3 +130,4 @@ pub use host_stub::*;
 mod host_default_error;
 #[cfg(any(test, feature = "test_utils"))]
 pub use host_default_error::*;
+use kitsune_p2p_types::metrics::MetricRecord;
