+++ crates/kitsune_p2p/timestamp/Cargo.toml
@@ -10,8 +10,6 @@ keywords = [ "holochain", "holo", "p2p", "dht", "networking" ]
 categories = [ "network-programming" ]
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 derive_more = "0.99"
 serde = { version = "1.0", features = ["derive"] }
