+++ crates/kitsune_p2p/types/Cargo.toml
@@ -20,6 +20,7 @@ kitsune_p2p_dht = { version = "^0.3.0-beta-dev.3", path = "../dht" }
 kitsune_p2p_dht_arc = { version = "^0.3.0-beta-dev.3", path = "../dht_arc" }
 kitsune_p2p_bin_data = { version = "^0.3.0-beta-dev.4", path = "../bin_data" }
 kitsune_p2p_block = { version = "^0.3.0-beta-dev.4", path = "../block" }
+kitsune_p2p_timestamp = { version = "0.2.0", path = "../timestamp" }
 lru = "0.8.1"
 mockall = { version = "0.11.3", optional = true }
 nanoid = "0.3"
@@ -36,7 +37,6 @@ shrinkwraprs = "0.3.0"
 sysinfo = "0.27"
 thiserror = "1.0.22"
 tokio = { version = "1.27", features = [ "full" ] }
-tokio-stream = { version = "0.1", features = [ "sync", "net" ] }
 url = "2"
 url2 = "0.0.6"
 webpki = "0.22.0"
